Remove configuration of HAVE_PUTENV.
The Python includes files for Windows define HAVE_PUTENV to be
empty. GnuCash under MINGW (only) defines it to be 1. This causes
redefinition errors when building the Python bindings on Windows.
HAVE_PUTENV is not used in the GnuCash code base, so the best
solution seems to be to just get rid of it.